# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "My API",
    "description": "Description",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/getorpost": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "textarea": {
                  "example": "any"
                },
                "id": {
                  "example": "any"
                },
                "rating": {
                  "example": "any"
                },
                "dateData": {
                  "example": "any"
                },
                "username": {
                  "example": "any"
                },
                "link": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/post": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "method": {
                  "example": "any"
                },
                "id": {
                  "example": "any"
                },
                "dateData": {
                  "example": "any"
                },
                "link": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "type": "string"
          },
          {
            "name": "id",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  }
}